Eliminating unnecessary public members (#345)
