+++ Directory.Packages.props
@@ -8,6 +8,7 @@
     <PackageVersion Include="Microsoft.Build.Framework" Version="17.3.1" />
     <PackageVersion Include="Microsoft.Build.Utilities.Core" Version="17.3.1" />
     <PackageVersion Include="Microsoft.Build" Version="17.3.1" />
+    <PackageVersion Include="Microsoft.CodeAnalysis.PublicApiAnalyzers" Version="3.3.4" />
     <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.4.1" />
     <PackageVersion Include="Microsoft.VisualStudioEng.MicroBuild.Core" Version="1.0.0" />
     <PackageVersion Include="Nerdbank.GitVersioning" Version="3.5.109" />
