+++ Microsoft.NET.Build.Containers/ContainerHelpers.cs
@@ -1,31 +1,20 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.NET.Build.Containers;
-
+#if NETFRAMEWORK
 using System;
+#endif
+#if NET
 using System.Diagnostics.CodeAnalysis;
+#endif
 using System.Text.RegularExpressions;
-using static ReferenceParser;
 
-internal sealed record Label(string name, string value);
-
-// Explicitly lowercase to ease parsing - the incoming values are
-// lowercased by spec
-public enum PortType
-{
-    tcp,
-    udp
-}
-
-public record Port(int number, PortType type);
-
-
+namespace Microsoft.NET.Build.Containers;
 public static class ContainerHelpers
 {
-    public const string HostObjectUser = "SDK_CONTAINER_REGISTRY_UNAME";
-    
-    public const string HostObjectPass = "SDK_CONTAINER_REGISTRY_PWORD";
+    internal const string HostObjectUser = "SDK_CONTAINER_REGISTRY_UNAME";
+
+    internal const string HostObjectPass = "SDK_CONTAINER_REGISTRY_PWORD";
 
     /// <summary>
     /// Matches an environment variable name - must start with a letter or underscore, and can only contain letters, numbers, and underscores.
@@ -37,7 +26,7 @@ public static class ContainerHelpers
     /// that have no registry component.
     /// See <see href="https://github.com/distribution/distribution/blob/78b9c98c5c31c30d74f9acb7d96f98552f2cf78f/reference/normalize.go">normalize.go</see>.
     /// </summary>
-    public const string DefaultRegistry = "docker.io";
+    internal const string DefaultRegistry = "docker.io";
 
     /// <summary>
     /// Matches if the string is not lowercase or numeric, or ., _, or -.
@@ -45,12 +34,101 @@ public static class ContainerHelpers
     /// <remarks>Technically the period should be allowed as well, but due to inconsistent support between cloud providers we're removing it.</remarks>
     private static Regex imageNameCharacters = new Regex(@"[^a-z0-9_\-/]");
 
+
+    /// <summary>
+    /// The enum contains possible error reasons during port parsing using <see cref="TryParsePort(string, out Port?, out ParsePortError?)"/> or <see cref="TryParsePort(string?, string?, out Port?, out ParsePortError?)"/>.
+    /// </summary>
+    [Flags]
+    public enum ParsePortError
+    {
+        MissingPortNumber,
+        InvalidPortNumber,
+        InvalidPortType,
+        UnknownPortFormat
+    }
+
+    /// <summary>
+    /// Tries to parse the port from <paramref name="portNumber"/> and <paramref name="portType"/>.
+    /// </summary>
+    /// <param name="portNumber">The port number to parse.</param>
+    /// <param name="portType">The port type to parse (tcp or udp).</param>
+    /// <param name="port">Parsed port.</param>
+    /// <param name="error">The error occurred during parsing. Only returned when method returns <see langword=""="false"/>.</param>
+    /// <returns><see langword=""="true"/> when port was successfully parsed, <see langword=""="false"/> otherwise.</returns>
+    public static bool TryParsePort(string? portNumber, string? portType, [NotNullWhen(true)] out Port? port, [NotNullWhen(false)] out ParsePortError? error)
+    {
+        var portNo = 0;
+        error = null;
+        if (String.IsNullOrEmpty(portNumber))
+        {
+            error = ParsePortError.MissingPortNumber;
+        }
+        else if (!int.TryParse(portNumber, out portNo))
+        {
+            error = ParsePortError.InvalidPortNumber;
+        }
+
+        if (!Enum.TryParse<PortType>(portType, out PortType t))
+        {
+            if (portType is not null)
+            {
+                error = (error ?? ParsePortError.InvalidPortType) | ParsePortError.InvalidPortType;
+            }
+            else
+            {
+                t = PortType.tcp;
+            }
+        }
+
+        if (error is null)
+        {
+            port = new Port(portNo, t);
+            return true;
+        }
+        else
+        {
+            port = null;
+            return false;
+        }
+
+    }
+
+    /// <summary>
+    /// Tries to parse the port from <paramref name="input"/>.
+    /// </summary>
+    /// <param name="input">The port number to parse. Expected formats are: port number as int value, or value in format 'port number/port type' where
+    /// port type can be tcp or udp. If the port type is not present, it is assumed to be tcp.</param>
+    /// <param name="port">Parsed port.</param>
+    /// <param name="error">The error occurred during parsing. Only returned when method returns <see langword="false"/>.</param>
+    /// <returns><see langword="true"/> when port was successfully parsed, <see langword="false"/> otherwise.</returns>
+    public static bool TryParsePort(string input, [NotNullWhen(true)] out Port? port, [NotNullWhen(false)] out ParsePortError? error)
+    {
+        var parts = input.Split('/');
+        if (parts.Length == 2)
+        {
+            string portNumber = parts[0];
+            string type = parts[1];
+            return TryParsePort(portNumber, type, out port, out error);
+        }
+        else if (parts.Length == 1)
+        {
+            string portNum = parts[0];
+            return TryParsePort(portNum, null, out port, out error);
+        }
+        else
+        {
+            error = ParsePortError.UnknownPortFormat;
+            port = null;
+            return false;
+        }
+    }
+
     /// <summary>
     /// Ensures the given registry is valid.
     /// </summary>
     /// <param name="registryName"></param>
     /// <returns></returns>
-    public static bool IsValidRegistry(string registryName) => AnchoredDomainRegexp.IsMatch(registryName);
+    internal static bool IsValidRegistry(string registryName) => ReferenceParser.AnchoredDomainRegexp.IsMatch(registryName);
 
     /// <summary>
     /// Ensures the given image name is valid.
@@ -58,9 +136,9 @@ public static class ContainerHelpers
     /// </summary>
     /// <param name="imageName"></param>
     /// <returns></returns>
-    public static bool IsValidImageName(string imageName)
+    internal static bool IsValidImageName(string imageName)
     {
-        return anchoredNameRegexp.IsMatch(imageName);
+        return ReferenceParser.anchoredNameRegexp.IsMatch(imageName);
     }
 
     /// <summary>
@@ -69,16 +147,16 @@ public static class ContainerHelpers
     /// </summary>
     /// <param name="imageTag"></param>
     /// <returns></returns>
-    public static bool IsValidImageTag(string imageTag)
+    internal static bool IsValidImageTag(string imageTag)
     {
-        return anchoredTagRegexp.IsMatch(imageTag);
+        return ReferenceParser.anchoredTagRegexp.IsMatch(imageTag);
     }
 
     /// <summary>
     /// Given an already-validated registry domain, this is our hueristic to determine what HTTP protocol should be used to interact with it.
     /// This is primarily for testing - in the real world almost all usage should be through HTTPS!
     /// </summary>
-    public static Uri TryExpandRegistryToUri(string alreadyValidatedDomain)
+    internal static Uri TryExpandRegistryToUri(string alreadyValidatedDomain)
     {
         var prefix = alreadyValidatedDomain.StartsWith("localhost", StringComparison.Ordinal) ? "http" : "https";
         return new Uri($"{prefix}://{alreadyValidatedDomain}");
@@ -89,7 +167,7 @@ public static class ContainerHelpers
     /// </summary>
     /// <param name="envVar"></param>
     /// <returns></returns>
-    public static bool IsValidEnvironmentVariable(string envVar)
+    internal static bool IsValidEnvironmentVariable(string envVar)
     {
         return envVarRegex.IsMatch(envVar);
     }
@@ -109,7 +187,7 @@ public static class ContainerHelpers
     /// <param name="containerTag"></param>
     /// <param name="containerDigest"></param>
     /// <returns>True if the parse was successful. When false is returned, all out vars are set to empty strings.</returns>
-    public static bool TryParseFullyQualifiedContainerName(string fullyQualifiedContainerName,
+    internal static bool TryParseFullyQualifiedContainerName(string fullyQualifiedContainerName,
                                                             [NotNullWhen(true)] out string? containerRegistry,
                                                             [NotNullWhen(true)] out string? containerName,
                                                             out string? containerTag, // tag is always optional - we can't guarantee anything here
@@ -118,7 +196,7 @@ public static class ContainerHelpers
     {
 
         /// if we don't have a reference at all, bail out
-        var referenceMatch = ReferenceRegexp.Match(fullyQualifiedContainerName);
+        var referenceMatch = ReferenceParser.ReferenceRegexp.Match(fullyQualifiedContainerName);
         if (referenceMatch is not { Success: true })
         {
             containerRegistry = null;
@@ -136,7 +214,7 @@ public static class ContainerHelpers
         // this will always be successful if the ReferenceRegexp matched, so it's safe to index into.
         var namePortion = referenceMatch.Groups[1].Value;
         // we try to decompose the reference name into registry and image name parts.
-        var nameMatch = anchoredNameRegexp.Match(namePortion);
+        var nameMatch = ReferenceParser.anchoredNameRegexp.Match(namePortion);
         if (nameMatch is { Success: true })
         {
             // the name regex has two groups:
@@ -183,7 +261,7 @@ public static class ContainerHelpers
     /// <summary>
     /// Checks if a given container image name adheres to the image name spec. If not, and recoverable, then normalizes invalid characters.
     /// </summary>
-    public static bool NormalizeImageName(string containerImageName,
+    internal static bool NormalizeImageName(string containerImageName,
                                          [NotNullWhen(false)] out string? normalizedImageName)
     {
         if (IsValidImageName(containerImageName))
@@ -199,75 +277,6 @@ public static class ContainerHelpers
             }
             var loweredImageName = containerImageName.ToLowerInvariant();
             normalizedImageName = imageNameCharacters.Replace(loweredImageName, "-");
-            return false;
-        }
-    }
-
-    [Flags]
-    public enum ParsePortError
-    {
-        MissingPortNumber,
-        InvalidPortNumber,
-        InvalidPortType,
-        UnknownPortFormat
-    }
-
-    public static bool TryParsePort(string? portNumber, string? portType, [NotNullWhen(true)] out Port? port, [NotNullWhen(false)] out ParsePortError? error)
-    {
-        var portNo = 0;
-        error = null;
-        if (String.IsNullOrEmpty(portNumber))
-        {
-            error = ParsePortError.MissingPortNumber;
-        }
-        else if (!int.TryParse(portNumber, out portNo))
-        {
-            error = ParsePortError.InvalidPortNumber;
-        }
-
-        if (!Enum.TryParse<PortType>(portType, out PortType t))
-        {
-            if (portType is not null)
-            {
-                error = (error ?? ParsePortError.InvalidPortType) | ParsePortError.InvalidPortType;
-            }
-            else
-            {
-                t = PortType.tcp;
-            }
-        }
-
-        if (error is null)
-        {
-            port = new Port(portNo, t);
-            return true;
-        }
-        else
-        {
-            port = null;
-            return false;
-        }
-
-    }
-
-    public static bool TryParsePort(string input, [NotNullWhen(true)] out Port? port, [NotNullWhen(false)] out ParsePortError? error)
-    {
-        var parts = input.Split('/');
-        if (parts.Length == 2)
-        {
-            string portNumber = parts[0];
-            string type = parts[1];
-            return TryParsePort(portNumber, type, out port, out error);
-        }
-        else if (parts.Length == 1)
-        {
-            string portNum = parts[0];
-            return TryParsePort(portNum, null, out port, out error);
-        }
-        else
-        {
-            error = ParsePortError.UnknownPortFormat;
-            port = null;
             return false;
         }
     }
