+++ Microsoft.NET.Build.Containers/Credentials/CredentialRetrievalException.cs
@@ -1,8 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
-
 namespace Microsoft.NET.Build.Containers.Credentials;
 
 internal sealed class CredentialRetrievalException : Exception
