+++ Microsoft.NET.Build.Containers/Globals.cs
@@ -1,0 +1,7 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Microsoft.NET.Build.Containers.UnitTests")]
+[assembly: InternalsVisibleTo("Microsoft.NET.Build.Containers.IntegrationTests")]
