+++ Microsoft.NET.Build.Containers/ImageReference.cs
@@ -6,7 +6,7 @@ namespace Microsoft.NET.Build.Containers;
 /// <summary>
 /// Represents a reference to a Docker image. A reference is made of a registry, a repository (aka the image name) and a tag.
 /// </summary>
-public readonly record struct ImageReference(Registry? Registry, string Repository, string Tag) {
+internal readonly record struct ImageReference(Registry? Registry, string Repository, string Tag) {
     public override string ToString()
     {
         if (Registry is {} reg) {
