+++ Microsoft.NET.Build.Containers/PublicAPI/net472/PublicAPI.Shipped.txt
@@ -1,0 +1,1 @@
+﻿#nullable enable
