+++ Microsoft.NET.Build.Containers/Tasks/CreateNewImage.Interface.cs
